dlr_wasm_interpreter_interop

Trait StoreTypedInvocationExt

pub trait StoreTypedInvocationExt<T: Config> {
    // Required methods
    fn func_alloc_typed<Params: InteropValueList, Returns: InteropValueList>(
        &mut self,
        hostcode: Hostcode,
    ) -> FuncAddr;
    unsafe fn invoke_simple_typed<Params: InteropValueList, Returns: InteropValueList>(
        &mut self,
        function: FuncAddr,
        params: Params,
    ) -> Result<Returns, RuntimeError>;
}

Required Methods

fn func_alloc_typed<Params: InteropValueList, Returns: InteropValueList>( &mut self, hostcode: Hostcode, ) -> FuncAddr

Allocates a new function with a statically known type signature with some host code.

This function is simply syntactic sugar for calling Store::func_alloc with statically know types.

unsafe fn invoke_simple_typed<Params: InteropValueList, Returns: InteropValueList>( &mut self, function: FuncAddr, params: Params, ) -> Result<Returns, RuntimeError>

Invokes a function without support for fuel or host functions but with a statically known type signature.

Safety

The caller has to guarantee that the given FuncAddr and any FuncAddr or ExternAddr values contained in the parameter values came from the current Store object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.